refactor: extract admin artifact projection
